adjust scripts
